updated for ViewModel

+++ UKParliament.CodeTest.Web/Controllers/PersonController.cs
@@ -1,8 +1,10 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using UKParliament.CodeTest.Data;
 using UKParliament.CodeTest.Services;
+using UKParliament.CodeTest.Web.ViewModels;
 
 namespace UKParliament.CodeTest.Web.Controllers
 {
@@ -11,56 +13,78 @@ namespace UKParliament.CodeTest.Web.Controllers
     public class PersonController : ControllerBase
     {
         private readonly IPersonService _personService;
+        private readonly IMapper _mapper;
 
-        public PersonController(IPersonService personService)
+        public PersonController(IPersonService personService, IMapper mapper)
         {
             _personService = personService;
+            _mapper = mapper;
         }
 
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Person>>> GetAll()
+        public async Task<ActionResult<IEnumerable<PersonViewModel>>> GetAll()
         {
-            return Ok(await _personService.GetAllAsync());
+            var persons = await _personService.GetAllAsync();
+            var personViewModels = _mapper.Map<IEnumerable<PersonViewModel>>(persons);
+            return Ok(personViewModels);
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<Person>> GetById(int id)
+        public async Task<ActionResult<PersonViewModel>> GetById(int id)
         {
             var person = await _personService.GetByIdAsync(id);
             if (person == null)
             {
                 return NotFound();
             }
-            return Ok(person);
+            var personViewModel = _mapper.Map<PersonViewModel>(person);
+            return Ok(personViewModel);
         }
 
         [HttpPost]
-        public async Task<IActionResult> Create([FromBody] Person person)
+        public async Task<IActionResult> Create([FromBody] PersonViewModel personViewModel)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
 
+            var person = _mapper.Map<Person>(personViewModel);
             await _personService.AddAsync(person);
-            return CreatedAtAction(nameof(GetById), new { id = person.Id }, person);
+            var createdPersonViewModel = _mapper.Map<PersonViewModel>(person);
+
+            return CreatedAtAction(nameof(GetById), new { id = person.Id }, createdPersonViewModel);
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Update(int id, [FromBody] Person person)
+        public async Task<IActionResult> Update(int id, [FromBody] PersonViewModel personViewModel)
         {
-            if (id != person.Id)
+            if (id != personViewModel.Id)
             {
                 return BadRequest();
             }
 
-            await _personService.UpdateAsync(person);
+            var existingPerson = await _personService.GetByIdAsync(id);
+            if (existingPerson == null)
+            {
+                return NotFound();
+            }
+
+            _mapper.Map(personViewModel, existingPerson);
+            await _personService.UpdateAsync(existingPerson);
+
             return NoContent();
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var existingPerson = await _personService.GetByIdAsync(id);
+            if (existingPerson == null)
+            {
+                return NotFound();
+            }
+
             await _personService.DeleteAsync(id);
             return NoContent();
         }
